app/requirements.txt
flask==2.1.1
Flask-Cors==3.0.10
flask_restful==0.3.9

mysql-connector-python
SQLAlchemy